Clippy changes
Source: https://github.com/msupply-foundation/open-msupply/commit/f22703eb222f5b2c8b779450f8639a447e0f0aba

+++ server/service/src/dashboard/item_count.rs
@@ -56,14 +56,14 @@ impl ItemCountServiceTrait for ItemServiceCount {
         let low_stock = item_stats
             .iter()
             .filter(|&i| (i.available_stock_on_hand > 0.0))
-            .map(|i| i.available_stock_on_hand as f64 / i.average_monthly_consumption)
+            .map(|i| i.available_stock_on_hand / i.average_monthly_consumption)
             .filter(|months_of_stock| *months_of_stock < low_stock_threshold as f64)
             .count() as i64;
 
         let more_than_six_months_stock = item_stats
             .iter()
             .filter(|&i| (i.available_stock_on_hand > 0.0))
-            .map(|i| i.available_stock_on_hand as f64 / i.average_monthly_consumption)
+            .map(|i| i.available_stock_on_hand / i.average_monthly_consumption)
             .filter(|months_of_stock| *months_of_stock > 6.0)
             .count() as i64;
